comments in dev mode script and also added exporting of the marks in export mode

+++ src/mapString.h
@@ -10,6 +10,6 @@ extern int originX;
 extern int originY;
 
 //Stores the exported map string that the map can be shows what you made
-extern const int mapString[];
+extern const float mapString[];
 
 #endif //RACINGGAME_MAPSTRING_H
